requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
Babel==2.7.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.6.16
cffi==1.12.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.0.3
decorator==4.4.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.2
django-celery==3.3.0
django-cors-headers==3.0.1
django-crispy-forms==1.7.2
django-db-reconnect==0.0.1
django-filter==2.1.0
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.3
djangorestframework==3.9.2
et-xmlfile==1.0.1
flower==0.9.3
future==0.17.1
httplib2==0.9.2
hyperlink==18.0.0
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
incremental==17.5.0
jdcal==1.4
Js2Py==0.66
kombu==3.0.37
lxml==4.3.2
Markdown==3.1
moviepy==1.0.0
Naked==0.1.31
numpy==1.16.2
odfpy==1.4.0
opencv-contrib-python==4.0.0.21
opencv-python==4.0.0.21
openpyxl==2.6.2
parsel==1.5.1
Pillow==6.1.0
proglog==0.1.9
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pycryptodome==3.9.0
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyjsparser==2.7.1
PyMySQL==0.9.3
PyMysqlDB==0.0.2
pyOpenSSL==19.0.0
python-scrapyd-api==2.1.2
pytz==2018.9
PyYAML==5.1
queuelib==1.5.0
redis==3.2.1
requests==2.22.0
Scrapy==1.6.0
selenium==3.141.0
service-identity==18.1.0
shellescape==3.4.1
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
tablib==0.13.0
tornado==5.1.1
tqdm==4.35.0
Twisted==18.9.0
tzlocal==2.0.0
urllib3==1.25.3
vine==1.3.0
w3lib==1.20.0
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
